Move implementation of partitioner
