+++ source/lac/la_sm_partitioner.inst.in
@@ -1,0 +1,62 @@
+// ---------------------------------------------------------------------
+//
+// Copyright (C) 2020 by the deal.II authors
+//
+// This file is part of the deal.II library.
+//
+// The deal.II library is free software; you can use it, redistribute
+// it, and/or modify it under the terms of the GNU Lesser General
+// Public License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+// The full text of the license can be found in the file LICENSE.md at
+// the top level directory of deal.II.
+//
+// ---------------------------------------------------------------------
+
+
+
+for (SCALAR : REAL_SCALARS)
+  {
+    namespace LinearAlgebra
+    \{
+      namespace SharedMPI
+      \{
+        template void
+        Partitioner::update_ghost_values_start(
+          SCALAR *                       data_this,
+          std::vector<SCALAR *> &        data_others,
+          dealii::AlignedVector<SCALAR> &buffer,
+          const unsigned int             communication_channel) const;
+
+        template void
+        Partitioner::update_ghost_values_finish(
+          SCALAR *                       data_this,
+          std::vector<SCALAR *> &        data_others,
+          dealii::AlignedVector<SCALAR> &buffer) const;
+
+        template void
+        Partitioner::update_ghost_values(
+          SCALAR *                       data_this,
+          std::vector<SCALAR *> &        data_others,
+          dealii::AlignedVector<SCALAR> &buffer) const;
+
+        template void
+        Partitioner::compress_start(
+          SCALAR *                       data_this,
+          std::vector<SCALAR *> &        data_others,
+          dealii::AlignedVector<SCALAR> &buffer,
+          const unsigned int             communication_channel) const;
+
+        template void
+        Partitioner::compress_finish(
+          SCALAR *                       data_this,
+          std::vector<SCALAR *> &        data_others,
+          dealii::AlignedVector<SCALAR> &buffer) const;
+
+        template void
+        Partitioner::compress(SCALAR *                       data_this,
+                              std::vector<SCALAR *> &        data_others,
+                              dealii::AlignedVector<SCALAR> &buffer) const;
+      \}
+    \}
+  }
